+++ HelloEarthTwo/Heroes.cs
@@ -1,22 +1,26 @@
-﻿using System;
+﻿using Newtonsoft.Json;
+using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
 namespace HelloEarthTwo
 {
-    public class Heroes
+    public class Heroes : Hero
     {
         public string CodeName { get; set; }
         public string Powers { get; set; }
-
         public string SecretId { get; set; }
         public string HomeWorld { get; set; }
+        public string TeamAffiliation { get; set; }
+        public string IsClone { get; set; }
 
-        public string IsClone { get; set; }  // may have to change type  to a string
+        /* public List<string> Powers { get; set; } = new();*/ // Can use this to turn our powers property to a list 
 
     }
+
 
 
 }
