+++ HelloEarthTwo/earthTwoHeroes.json
@@ -1,1 +1,1 @@
-{"codeName":"The Flash","powers":"Hyper Speed","secretId":"Jay Garreck","homeWorld":"Earth-2","isClone":"no"}
+{"codeName":"Superman","powers":"Strength","secretId":"Val-Zod","homeWorld":"Earth-2","teamAffiliation":"JSA","isClone":"no"}
