+++ HelloEarthTwo/multiverseHeroes.json
@@ -1,1 +1,1 @@
-{"codeName":"The Flash","powers":"Speed Force","secretId":"Barry Allen","homeWorld":"Earth-1","isClone":"no"}
+{"codeName":"Superman","powers":"flight","secretId":"Clark Kent","homeWorld":"Earth","teamAffiliation":"JLA","isClone":"no"}
